Added new initial create because the models have changed

+++ WebServer/DataContext/201510271614524_InitialCreate.cs
@@ -37,7 +37,7 @@ namespace WebServer.DataContext
                     {
                         Id = c.Int(nullable: false, identity: true),
                         IsConnected = c.Boolean(nullable: false),
-                        TimeStamp = c.Int(nullable: false),
+                        TimeStamp = c.Long(nullable: false),
                         SmartPlaneUser_Id = c.Int(),
                     })
                 .PrimaryKey(t => t.Id)
@@ -50,7 +50,7 @@ namespace WebServer.DataContext
                     {
                         Id = c.Int(nullable: false, identity: true),
                         Value = c.Int(nullable: false),
-                        TimeStamp = c.Int(nullable: false),
+                        TimeStamp = c.Long(nullable: false),
                         SmartPlaneUser_Id = c.Int(),
                     })
                 .PrimaryKey(t => t.Id)
@@ -63,7 +63,7 @@ namespace WebServer.DataContext
                     {
                         Id = c.Int(nullable: false, identity: true),
                         Value = c.Int(nullable: false),
-                        TimeStamp = c.Int(nullable: false),
+                        TimeStamp = c.Long(nullable: false),
                         SmartPlaneUser_Id = c.Int(),
                     })
                 .PrimaryKey(t => t.Id)
